update xadf to directly call functions during install
Previously it is trying to call itself to do the job, but then it
does not honor --seat option. Since --seat option sets working
variables, if we call another instance of the script to do the job
it would not honor --seat. If the function is directly called, then
it would also work with the same set of variables