Code improvements: use cp instead of cat for replicating templates
Also change tests on whether bash_functions and bash_aliases exist
from recipe.txt to xadfrc, as xadfrc is meant to be generated on
upgrade, but recipe.txt might contain user configurations.

For my future self, remember that while recipe.txt is configurable,
xadfrc is meant to be generated by xadf itself, not by user. There
is a reason why the content of xadfrc is standarized, but not the
content of recipe.txt