changes some variable names for future upgrades
xadf_https_repo is renamed to xadf_clone_url as it is the url used
to clone from in step 1-5 of installation. It is renamed as such so
that in the future we can modify its source url in order for future
user can customize from which repository they can download from.

xadf_ssh_repo is renamed to xadf_set_url as it is the url used
specifically after installation or during installation that alters
git config files, especially url of origin. In the future it can
be a short alias for git remote set-url origin URL.