update build_xadfrc() to just use one sed and one redirection
Previously it will source template-xadfrc with embedded here doc
notation, then piped to sed to remove all $HOME to literal '$HOME'.
However that was not efficient, and adds complication, as we have
to escape all '$' character, for example.

With this update, the file template-xadfrc is identical to xadfrc
except for the part where the value for xadfdir is parsed. In the
template-xadfrc it is replaced with dummy text XADFRC. Then func
build_xadfrc will just use one sed invocation to first replace
XADFDIR with variable $install_seat (therefore configurable by
passing '--seat DIR'), and then replace all $HOME with literal
'$HOME'.

This one will result in a truly reproducible xadfrc and is much
simpler to modify, as we don't have to escape any characters.