adds .profile so it works perfectly with i3 and dmenu_run
dmenu_run is peculiar, it sources .profile not .bashrc, which by
itself is not surprising, since it is not run from terminal.
However my bashrc is included in .profile, but dmenu_run fail to
do so. Therefore it fails to include my local bin to PATH.

Turns out it is because ubuntu defaults to sourcing .bashrc first
before setting up custom local path. After reversing the order,
dmenu_run works perfectly and is able to set my custom PATH, hence
the ability to run local executables from dmenu_run

Also I made .profile terser than the ubuntu default already, so why
not include it to trunk?