stack.sh: updates print to allow custom indices select
Previously it just prints the largest indices, however I think
it is also convenient to be able to access values of any indices
inside a stack, or so to peek the content.

The means of which it is done is simply by having an additional
argument parsing structure, that will parse arguments -i or --input
followed with an integer, then it will be passed as indices. Then
the original indices selector is being replaced with just a
variable, which contains the default value of -1, or any highest
indices.