xadfcommit: modify methods inside xadf_check_path()
Now it will check if path is provided already. If it isn't. then
there is no -p/--path flag issued. Then it will have to assume that
xadf is modified already. If that is the case, then it will get from
git status, and obtain path to xadf. If it still couldn't find the
path to xadf, then it will just scream.