combine updates from trunk to master
Changes summary:
- adds help text for double dash in xadf main script
- adds xadfcommit script, derived from xadfcommit()
- adds il-aid.sh, helper functions for my own use
- adds submodules: ohmyzsh and zsh_custom
- protects xadfrc for the cases where bash is not used, and separating some functions valid for zsh and bash to initrc
- updates nls to also support symlinked folders
- updates some dotfiles configurations

Note that in the future we might need to separate the content from the tool.
Currently we have them all in a single repository: the tools and the dotfiles content.
In the future I would like them to be separate, so I can focus to work on them separately.