adapts tmx with https://wiki.archlinux.org/title/Tmux
From the page, we get the following explanation:

The tmx script below implements this — the version here is slightly
modified to execute tmux new-window if 1 is its second parameter.
Invoked as tmx base_session_name [1], it launches the base session
if necessary. Otherwise a new "client" session linked to the base,
optionally add a new window and attach, setting it to kill itself
once it turns "zombie". Do not forget to make it executable.