tmx: uses a custom datetime hash to avoid conflict
The previous solution is ugly, as it is too long. I want a short
one that is not very likely to conflict (as if, I don't think I
would be having so many sessions attached to the base session that
some of the session name clashed.

So I just extract the datetime hash with md5sum, and take only the
first 4 letters as session id.