tmx: switches cleaning up to archwiki method
Previously it checks for any detached sessions names with a
sequence of 14 numbers (for YYYYMMDDHHMMSS). We have difficulties
to reliably change the format to [a-f0-9] with 4 letter strings, as
it will match any sessions that fall within that scope. Then I
realized that the archwiki version has a cleaner solution, by
making each attached sessions automatically exit (and therefore
declutter the sessions), instead of manually deleting inactive
sessions.

I will keep that behavior, makes the code a lot simpler to read.
But we will keep the current session naming convention (the first
four letters of unix epoch md5sum).