adds sourcing mechanism with updated recipe.txt format
Previously we directly sources recipe.txt format. However with this
iteration, we first parse recipe.txt file line by line:
1. First we strip off any comments
2. Then we strip off lines with leading spaces
3. Then we only put nonzero length lines to an array

The end result is an array containing all modules to be loaded. The
module list array is then used to load function and alias
definitions in:

$xadfmods/
  common/ # Checked first
  bash/   # If not found in common/ and is in bash
  zsh/    # If not found in common/ and is in zsh

When they are found, we uses src() to load them. Basically we can
now load "superfunctions", or functions split in multiple files,
but is contained in a single directory under directories described
above. For example stack superfunctions can be split into multiple
files under $xadfmods/bash/stack/ and we can split the file for
readability.

The next job is to modify default recipe.txt