adds zsh/ohmy.zsh submodule to account the use of ohmyzsh
This is a part of an effort to tackle task #7: to handle ohmyzsh
module. The solution is rather simple: than to have ohmyzsh dictate
how to manage my .zshrc, I take the control back by making .zshrc
managed by xadf, and ohmyzsh integration is added just as a xadf
module.