Reorganizes minimal required file locations in minimal-install
Our strictly required files has changed from:

.local/bin/xadf
.local/xadf/templates/bash_aliases
.local/xadf/templates/bash_functions
.local/xadf/templates/default-recipe.txt
.local/xadf/templates/template-xadfrc

To:

.local/bin/xadf
.local/xadf/initrc
.local/share/xadf/recipe.txt
.local/share/xadf/xadfrc

Therefore it is required for us to restructure our script so it
would download minimal install at the correct locations. Even if
it is not a minimal install, we still need the correct locations
to install default files such as recipe.txt and xadfrc.

Addresses #15:
- Modify xadf's xadf_build_xadfrc() function to reflect the change
- Modify xadf's xadf_build_recipe() function to reflect the change
- Modify xadf's xadf_curl_download() function to reflect the change
- Modify xadf's xadf_wget_download() function to reflect the change
- Modify xadf's xadf_minimal_test() function to reflect the change